Refactor tests structure for lists
